Create mainprogram9 - Long version.py
Since we change the data recording structure (I have mentioned in the previous README file), I change the loadData function in your code to run a few test.
You can try to test your code with the data in "Find Peak - Long test code" folder: "Pulse_20200831-110833_data.txt" and "Pulse_20200913-162824_data.txt"
I have tested it with your code, 1 is ok, one is so so (missing a few peak). The data row index is 3 and x row index is 0
If you want to test it with "vna_cont_spec_pulselike.txt" file, the data row index is 1 and the x row index is 0